Version 0.2.1 	- Settings incluye the configuracion de la base de datos para mysql. Esta comentada. Sigue la configuración de sqlite3. 	- Numero de Socio pasa a ser models.IntegerField, para que el primer superusuario pueda tener un valor negativo.

+++ usuarios/models.py
@@ -117,11 +117,10 @@ class CustomUser(AbstractUser):
                            null=True,
                            blank=True)
 
-    numero_socio = models.PositiveSmallIntegerField(
-        verbose_name='Número de Socio',
-        default=0,
-        null=True,
-        blank=True)
+    numero_socio = models.IntegerField(verbose_name='Número de Socio',
+                                       default=0,
+                                       null=True,
+                                       blank=True)
 
     profesion = models.CharField(verbose_name='Profesión',
                                  max_length=100,
